+++ scripts/macros.js
@@ -288,6 +288,8 @@ export {relentlessHunter} from './macros/2024/classFeatures/ranger/relentlessHun
 export {naturesVeil} from './macros/2024/classFeatures/ranger/naturesVeil.js';
 export {preciseHunter} from './macros/2024/classFeatures/ranger/preciseHunter.js';
 export {foeSlayer} from './macros/2024/classFeatures/ranger/foeSlayer.js';
+//Beast Master
+export {primalCompanion} from './macros/2024/classFeatures/ranger/beastMaster/primalCompanion.js';
 //Rogue
 export {sneakAttack} from './macros/2024/classFeatures/rogue/sneakAttack.js';
 export {cunningAction} from './macros/2024/classFeatures/rogue/cunningAction.js';